BAO workload utility overhaul
This fixes many bugs and introduces a new reset mode which deletes all
previous experience from the BAO model.

However, the utility still lacks a proper way of executing a workload
"as intended", due to limitations in BAO:
Currently, BAO can only learn from regulare queries, i.e. queries of the
form `SELECT ... FROM ...` But we are not interested in the actual output
of the query, but in the plans generated by the optimizer. This
information is retrieved by running queries as `EXPLAIN ANALYZE SELECT
... FROM ...`. If the query is executed that way, BAO does not learn
anything.