Add raw result files for pg-bao experiments
Beware that these results were produce on a machine different from the
other experiments! This makes them uncomparable to the results in the
ues and pg-aqo folders (at least on an absolute scale). Most
importantly, the results on pg-bao were achieved on a machine were both
the database, as well as the database server fully reside on SSD
storage.