Fix race conditions in BAO server start
The race condition appeared when a BAO server was stopped quickly after
being started and the proper exit handler had not been set up yet. This
led to dangling socket connections on the port a future BAO server
instance was supposed to be using. In total, this prevented the server
from being started/stopped properly and thus made clearing the learned
data impossible.

The workaround for this problem is to wait for a small amount of time
after starting the server, to ensure that the shutdown handler has been
initialized properly.